rename mtools, gamma distribution prod ready
