+++ src/main/java/rapaio/core/distributions/ChiSquare.java
@@ -24,7 +24,7 @@
 
 package rapaio.core.distributions;
 
-import rapaio.math.MathTools;
+import rapaio.math.MTools;
 import rapaio.core.RandomSource;
 import rapaio.sys.WS;
 
@@ -66,7 +66,7 @@ public class ChiSquare implements Distribution {
     public double pdf(double x) {
         if (x < 0.0)
             return 0;
-        double logGamma = MathTools.lnGamma(df / 2.0);
+        double logGamma = MTools.lnGamma(df / 2.0);
         return Math.exp((df / 2.0 - 1.0) * Math.log(x / 2.0) - x / 2.0 - logGamma) / 2.0;
     }
 
@@ -74,7 +74,7 @@ public class ChiSquare implements Distribution {
     public double cdf(double x) {
         if (x < 0.0 || df < 1.0)
             return 0.0;
-        return MathTools.incompleteGamma(df / 2.0, x / 2.0);
+        return MTools.incompleteGamma(df / 2.0, x / 2.0);
     }
 
     @Override
