+++ src/main/java/rapaio/core/stat/Sum.java
@@ -24,7 +24,7 @@
 
 package rapaio.core.stat;
 
-import rapaio.math.MathTools;
+import rapaio.math.MTools;
 import rapaio.data.Var;
 import rapaio.printer.Printable;
 
@@ -33,7 +33,7 @@ import static rapaio.sys.WS.formatFlex;
 /**
  * Computes the sum of elements for a {@link rapaio.data.Var} of values.
  * <p>
- * Ignore invalid numeric values. See {@link MathTools#validNumber(double)}.
+ * Ignore invalid numeric values. See {@link MTools#validNumber(double)}.
  * <p>
  * User: <a href="mailto:padreati@yahoo.com">Aurelian Tutuianu</a>
  */
